+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 # 1.1.0 (not yet released)
 
+- [issue #1] Don't blow up on `<tbody>` or `<thead>`.
 - [issue #2] Fix `<li><p>...` case where the bullet would get dropped.
 - [issue #4] Don't drop internal links (i.e. to `#my-anchor` in the same doc).
 - [issue #3] fix `<code>` handling